Task spec:plugins
I had some sort of conflict when using rspec-rails and restful_authentication in a project. restful_authentication has generators' templates with names like spec/controllers/access_control_spec.rb and other of the same kind. So, rspec-rails tries to execute them when the task spec:plugins is run. I changed this task to avoid it and also to allow me to run specs of a specific plugin.